+++ examples/creando-procesos/creando-procesos.sh
@@ -1,6 +1,10 @@
 # The spawned programs return output that is the same
 # as if we had run them directly from the command-line.
-$ go run spawning-processes.go 
+# El proceso creado regresa la salida de la misma
+# forma que si lo hubieramos corrido directamente desde
+# la línea de comandos.
+
+$ go run creando-procesos.go 
 > date
 Wed Oct 10 09:53:11 PDT 2012
 
@@ -10,4 +14,4 @@ hello grep
 > ls -a -l -h
 drwxr-xr-x  4 mark 136B Oct 3 16:29 .
 drwxr-xr-x 91 mark 3.0K Oct 3 12:50 ..
--rw-r--r--  1 mark 1.3K Oct 3 16:28 spawning-processes.go
+-rw-r--r--  1 mark 1.3K Oct 3 16:28 creando-procesos.go
