+++ public/execing-processes
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta http-eqiv="content-type" content="text/html;charset=utf-8">
-    <title>Go con Ejemplos: Exec'ing Processes</title>
+    <title>Go con Ejemplos: Execing Processes</title>
     <link rel=stylesheet href="site.css">
   </head>
   <script>
@@ -17,18 +17,18 @@
   </script>
   <body>
     <div class="example" id="execing-processes">
-      <h2><a href="./">Go con Ejemplos</a>: Exec'ing Processes</h2>
+      <h2><a href="./">Go con Ejemplos</a>: Execing Processes</h2>
       
       <table>
         
         <tr>
           <td class="docs">
             <p>In the previous example we looked at
-<a>spawning external processes</a>. We
+<a href="spawning-processes">spawning external processes</a>. We
 do this when we need an external process accessible to
 a running Go process. Sometimes we just want to
 completely replace the current Go process with another
-(perhaps non-Go) one. To do this we’ll use Go’s
+(perhaps non-Go) one. To do this we&rsquo;ll use Go&rsquo;s
 implementation of the classic
 <a href="http://en.wikipedia.org/wiki/Exec_(operating_system)"><code>exec</code></a>
 function.</p>
@@ -80,9 +80,9 @@ function.</p>
         
         <tr>
           <td class="docs">
-            <p>For our example we’ll exec <code>ls</code>. Go requires an
+            <p>For our example we&rsquo;ll exec <code>ls</code>. Go requires an
 absolute path to the binary we want to execute, so
-we’ll use <code>exec.LookPath</code> to find it (probably
+we&rsquo;ll use <code>exec.LookPath</code> to find it (probably
 <code>/bin/ls</code>).</p>
 
           </td>
@@ -100,7 +100,7 @@ we’ll use <code>exec.LookPath</code> to find it (probably
         <tr>
           <td class="docs">
             <p><code>Exec</code> requires arguments in slice form (as
-apposed to one big string). We’ll give <code>ls</code> a few
+apposed to one big string). We&rsquo;ll give <code>ls</code> a few
 common arguments. Note that the first argument should
 be the program name.</p>
 
@@ -115,7 +115,7 @@ be the program name.</p>
         
         <tr>
           <td class="docs">
-            <p><code>Exec</code> also needs a set of <a>environment variables</a>
+            <p><code>Exec</code> also needs a set of <a href="environment-variables">environment variables</a>
 to use. Here we just provide our current
 environment.</p>
 
@@ -130,10 +130,10 @@ environment.</p>
         
         <tr>
           <td class="docs">
-            <p>Here’s the actual <code>os.Exec</code> call. If this call is
+            <p>Here&rsquo;s the actual <code>os.Exec</code> call. If this call is
 successful, the execution of our process will end
 here and be replaced by the <code>/bin/ls -a -l -h</code>
-process. If there is an error we’ll get a return
+process. If there is an error we&rsquo;ll get a return
 value.</p>
 
           </td>
@@ -173,8 +173,8 @@ value.</p>
         <tr>
           <td class="docs">
             <p>Note that Go does not offer a classic Unix <code>fork</code>
-function. Usually this isn’t an issue though, since
-starting goroutines, spawning processes, and exec’ing
+function. Usually this isn&rsquo;t an issue though, since
+starting goroutines, spawning processes, and exec&rsquo;ing
 processes covers most use cases for <code>fork</code>.</p>
 
           </td>
