+++ public/index.html
@@ -157,9 +157,9 @@
       
         <li><a href="variables-de-entorno">Variables de entorno</a></li>
       
-        <li><a href="spawning-processes">Spawning Processes</a></li>
+        <li><a href="creando-procesos">Creando procesos</a></li>
       
-        <li><a href="exec-ing-processes">Exec'ing Processes</a></li>
+        <li><a href="execing-processes">Execing Processes</a></li>
       
         <li><a href="signals">Signals</a></li>
