+++ public/variables-de-entorno
@@ -175,7 +175,7 @@ imprime en pantalla.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="spawning-processes">Spawning Processes</a>.
+        Siguiente ejemplo: <a href="creando-procesos">Creando procesos</a>.
       </p>
       
       <p class="footer">
